feat(accordion): added descriptive accordion
